refactor: move ui.model.Product to model.Product
