+++ app/src/androidTest/java/nextstep/shoppingcart/ui/product/list/ProductListItemTest.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.shoppingcart.ui.model.Product
+import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.product.list.component.ProductListItem
 import org.junit.Before
 import org.junit.Rule
