+++ app/src/main/java/nextstep/shoppingcart/ui/product/detail/ProductDetailActivity.kt
@@ -4,7 +4,7 @@ import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.shoppingcart.ui.model.Product
+import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 @Suppress("DEPRECATION")
