+++ app/src/main/java/nextstep/shoppingcart/ui/product/detail/ProductDetailScreen.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import nextstep.shoppingcart.R
-import nextstep.shoppingcart.ui.model.Product
+import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.theme.Black10
 import nextstep.shoppingcart.ui.theme.Blue50
 import nextstep.shoppingcart.ui.theme.Gray10
